Perfomance improvment; refactoring
Perfomance improvment:
__generate_urls in src/parser.py - use a generator instead of returning a list

Refactoring:
update gitignore - added *cache (for ignore ruff, pytest & other cache)
src/constants.py - changed SEMAPHORE_MAX_LIMIT
renaming src/exceptions/constants.py to messages.py
