+++ .gitignore
@@ -1,4 +1,4 @@
 .vscode
 __pycache__
-.pytest_cache
+*cache
 /parser-output
